AbstractEnderDiskItem and update FTBTeamsCompat
Pulled shared code between EnderDiskItem and EnderFluidDiskItem into AbstractEnderDiskItem
Updated FTBTeamsCompat to match AbstractEnderDiskItem
Updated lang files with fluid disks
Converted som Component.literal tooltips to Component.translatable
