+++ common/src/main/resources/assets/enderdrives/lang/en_us.json
@@ -1,57 +1,59 @@
 {
-  "key.categories.enderdrives": "EnderDrives",
-  "item.enderdrives.ender_storage_component_1k": "1k Ender Storage Component",
-  "item.enderdrives.ender_storage_component_4k": "4k Ender Storage Component",
-  "item.enderdrives.ender_storage_component_16k": "16k Ender Storage Component",
-  "item.enderdrives.ender_storage_component_64k": "64k Ender Storage Component",
-  "item.enderdrives.ender_storage_component_256k": "256k Ender Storage Component",
-  "item.enderdrives.ender_disk_1k": "1k Ender Item Storage Cell",
-  "item.enderdrives.ender_disk_4k": "4k Ender Item Storage Cell",
-  "item.enderdrives.ender_disk_16k": "16k Ender Item Storage Cell",
-  "item.enderdrives.ender_disk_64k": "64k Ender Item Storage Cell",
-  "item.enderdrives.ender_disk_256k": "256k Ender Item Storage Cell",
-  "item.enderdrives.ender_disk_creative": "Creative Ender Item Storage Cell",
-  "item.enderdrives.ender_fluid_storage_component_1k": "1k Ender Fluid Storage Component",
-  "item.enderdrives.ender_fluid_storage_component_4k": "4k Ender Fluid Storage Component",
-  "item.enderdrives.ender_fluid_storage_component_16k": "16k Ender Fluid Storage Component",
-  "item.enderdrives.ender_fluid_storage_component_64k": "64k Ender Fluid Storage Component",
-  "item.enderdrives.ender_fluid_storage_component_256k": "256k Ender Fluid Storage Component",
-  "item.enderdrives.ender_fluid_disk_1k": "1k Ender Fluid Storage Cell",
-  "item.enderdrives.ender_fluid_disk_4k": "4k Ender Fluid Storage Cell",
-  "item.enderdrives.ender_fluid_disk_16k": "16k Ender Fluid Storage Cell",
-  "item.enderdrives.ender_fluid_disk_64k": "64k Ender Fluid Storage Cell",
-  "item.enderdrives.ender_fluid_disk_256k": "256k Ender Fluid Storage Cell",
-  "item.enderdrives.ender_fluid_disk_creative": "Creative Ender Fluid Storage Cell",
-  "item.enderdrives.tape_disk": "Tape Disk Item Storage Cell",
   "enderdrives.transfer_mode.bidirectional": "Bidirectional",
   "enderdrives.transfer_mode.input": "Input Only",
   "enderdrives.transfer_mode.output": "Output Only",
-  "screen.enderdrives.frequency_selector": "Ender Frequency Selector",
+  "item.enderdrives.ender_disk_16k": "16k Ender Item Storage Cell",
+  "item.enderdrives.ender_disk_1k": "1k Ender Item Storage Cell",
+  "item.enderdrives.ender_disk_256k": "256k Ender Item Storage Cell",
+  "item.enderdrives.ender_disk_4k": "4k Ender Item Storage Cell",
+  "item.enderdrives.ender_disk_64k": "64k Ender Item Storage Cell",
+  "item.enderdrives.ender_disk_creative": "Creative Ender Item Storage Cell",
+  "item.enderdrives.ender_fluid_disk_16k": "16k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_1k": "1k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_256k": "256k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_4k": "4k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_64k": "64k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_creative": "Creative Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_storage_component_16k": "16k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_1k": "1k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_256k": "256k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_4k": "4k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_64k": "64k Ender Fluid Storage Component",
+  "item.enderdrives.ender_storage_component_16k": "16k Ender Storage Component",
+  "item.enderdrives.ender_storage_component_1k": "1k Ender Storage Component",
+  "item.enderdrives.ender_storage_component_256k": "256k Ender Storage Component",
+  "item.enderdrives.ender_storage_component_4k": "4k Ender Storage Component",
+  "item.enderdrives.ender_storage_component_64k": "64k Ender Storage Component",
+  "item.enderdrives.tape_disk": "Tape Disk Item Storage Cell",
+  "key.categories.enderdrives": "EnderDrives",
   "screen.enderdrives.frequency": "EnderDisk Frequency",
-  "screen.enderdrives.frequency_placeholder": "Frequency...",
   "screen.enderdrives.frequency_label": "Frequency",
-  "screen.enderdrives.toggle_scope": "Click to toggle scope",
-  "screen.enderdrives.tooltip.plus_256": "+256",
-  "screen.enderdrives.tooltip.plus_16": "+16",
-  "screen.enderdrives.tooltip.plus_1": "+1",
-  "screen.enderdrives.tooltip.minus_256": "-256",
-  "screen.enderdrives.tooltip.minus_16": "-16",
-  "screen.enderdrives.tooltip.minus_1": "-1",
+  "screen.enderdrives.frequency_placeholder": "Frequency...",
+  "screen.enderdrives.frequency_selector": "Ender Frequency Selector",
   "screen.enderdrives.scope.global": "Global",
   "screen.enderdrives.scope.private": "Private",
   "screen.enderdrives.scope.team": "Team",
-  "tooltip.enderdrives.types": "%s§7 of §9%s§7 Types",
+  "screen.enderdrives.toggle_scope": "Click to toggle scope",
+  "screen.enderdrives.tooltip.minus_1": "-1",
+  "screen.enderdrives.tooltip.minus_16": "-16",
+  "screen.enderdrives.tooltip.minus_256": "-256",
+  "screen.enderdrives.tooltip.plus_1": "+1",
+  "screen.enderdrives.tooltip.plus_16": "+16",
+  "screen.enderdrives.tooltip.plus_256": "+256",
+  "tooltip.enderdrives.disabled": "§cThis item is disabled on the server.",
+  "tooltip.enderdrives.enderdisk.disabled": "§cThis EnderDisk is disabled on the server.",
   "tooltip.enderdrives.fluid_types": "%s§7 of §9%s§7 Fluid Types",
+  "tooltip.enderdrives.fluidenderdisk.disabled": "§cThis FluidEnderDisk is disabled on the server.",
   "tooltip.enderdrives.frequency": "§7Frequency: §e%s",
+  "tooltip.enderdrives.mode.bidirectional": "§7Mode: §aBidirectional",
   "tooltip.enderdrives.mode.input": "§7Mode: §9Input Only",
   "tooltip.enderdrives.mode.output": "§7Mode: §cOutput Only",
-  "tooltip.enderdrives.mode.bidirectional": "§7Mode: §aBidirectional",
+  "tooltip.enderdrives.partitioned": "§7Partitioned: §f%s item%s",
+  "tooltip.enderdrives.scope.global": "§7Global",
   "tooltip.enderdrives.scope.private": "§7Private: §e%s",
   "tooltip.enderdrives.scope.team": "§7Team: §e%s",
-  "tooltip.enderdrives.scope.global": "§7Global",
-  "tooltip.enderdrives.partitioned": "§7Partitioned: §f%s item%s",
-  "tooltip.enderdrives.unknown": "Unknown",
+  "tooltip.enderdrives.tape.bytes": "%1$s of %2$s Bytes Used",
   "tooltip.enderdrives.tape.types": "%1$s of %2$s Types",
-  "tooltip.enderdrives.tape.bytes": "%1$s of %2$s Bytes Used"
-
+  "tooltip.enderdrives.types": "%s§7 of §9%s§7 Types",
+  "tooltip.enderdrives.unknown": "Unknown"
 }
