+++ common/src/main/resources/assets/enderdrives/lang/ja_jp.json
@@ -1,43 +1,59 @@
 {
-  "key.categories.enderdrives": "EnderDrives",
-  "item.enderdrives.ender_storage_component_1k": "1K エンダーストレージコンポーネント",
-  "item.enderdrives.ender_storage_component_4k": "4K エンダーストレージコンポーネント",
-  "item.enderdrives.ender_storage_component_16k": "16K エンダーストレージコンポーネント",
-  "item.enderdrives.ender_storage_component_64k": "64K エンダーストレージコンポーネント",
-  "item.enderdrives.ender_storage_component_256k": "256K エンダーストレージコンポーネント",
-  "item.enderdrives.ender_disk_1k": "1k エンダーアイテムストレージセル",
-  "item.enderdrives.ender_disk_4k": "4k エンダーアイテムストレージセル",
-  "item.enderdrives.ender_disk_16k": "16k エンダーアイテムストレージセル",
-  "item.enderdrives.ender_disk_64k": "64k エンダーアイテムストレージセル",
-  "item.enderdrives.ender_disk_256k": "256k エンダーアイテムストレージセル",
-  "item.enderdrives.ender_disk_creative": "クリエイティブエンダーアイテムストレージセル",
-  "item.enderdrives.tape_disk": "テープディスクアイテムストレージセル",
   "enderdrives.transfer_mode.bidirectional": "双方向",
   "enderdrives.transfer_mode.input": "入力のみ",
   "enderdrives.transfer_mode.output": "出力のみ",
-  "screen.enderdrives.frequency_selector": "エンダー周波数セレクター",
+  "item.enderdrives.ender_disk_16k": "16k エンダーアイテムストレージセル",
+  "item.enderdrives.ender_disk_1k": "1k エンダーアイテムストレージセル",
+  "item.enderdrives.ender_disk_256k": "256k エンダーアイテムストレージセル",
+  "item.enderdrives.ender_disk_4k": "4k エンダーアイテムストレージセル",
+  "item.enderdrives.ender_disk_64k": "64k エンダーアイテムストレージセル",
+  "item.enderdrives.ender_disk_creative": "クリエイティブエンダーアイテムストレージセル",
+  "item.enderdrives.ender_fluid_disk_16k": "16k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_1k": "1k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_256k": "256k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_4k": "4k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_64k": "64k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_creative": "Creative Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_storage_component_16k": "16k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_1k": "1k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_256k": "256k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_4k": "4k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_64k": "64k Ender Fluid Storage Component",
+  "item.enderdrives.ender_storage_component_16k": "16K エンダーストレージコンポーネント",
+  "item.enderdrives.ender_storage_component_1k": "1K エンダーストレージコンポーネント",
+  "item.enderdrives.ender_storage_component_256k": "256K エンダーストレージコンポーネント",
+  "item.enderdrives.ender_storage_component_4k": "4K エンダーストレージコンポーネント",
+  "item.enderdrives.ender_storage_component_64k": "64K エンダーストレージコンポーネント",
+  "item.enderdrives.tape_disk": "テープディスクアイテムストレージセル",
+  "key.categories.enderdrives": "EnderDrives",
   "screen.enderdrives.frequency": "エンダーディスク周波数",
-  "screen.enderdrives.frequency_placeholder": "周波数...",
   "screen.enderdrives.frequency_label": "周波数",
-  "screen.enderdrives.toggle_scope": "クリックして範囲を切り替える",
-  "screen.enderdrives.tooltip.plus_256": "+256",
-  "screen.enderdrives.tooltip.plus_16": "+16",
-  "screen.enderdrives.tooltip.plus_1": "+1",
-  "screen.enderdrives.tooltip.minus_256": "-256",
-  "screen.enderdrives.tooltip.minus_16": "-16",
-  "screen.enderdrives.tooltip.minus_1": "-1",
+  "screen.enderdrives.frequency_placeholder": "周波数...",
+  "screen.enderdrives.frequency_selector": "エンダー周波数セレクター",
   "screen.enderdrives.scope.global": "グローバル",
   "screen.enderdrives.scope.private": "プライベート",
   "screen.enderdrives.scope.team": "チーム",
-  "tooltip.enderdrives.types": "%s§7 の §9%s§7 タイプ",
+  "screen.enderdrives.toggle_scope": "クリックして範囲を切り替える",
+  "screen.enderdrives.tooltip.minus_1": "-1",
+  "screen.enderdrives.tooltip.minus_16": "-16",
+  "screen.enderdrives.tooltip.minus_256": "-256",
+  "screen.enderdrives.tooltip.plus_1": "+1",
+  "screen.enderdrives.tooltip.plus_16": "+16",
+  "screen.enderdrives.tooltip.plus_256": "+256",
+  "tooltip.enderdrives.disabled": "§cThis item is disabled on the server.",
+  "tooltip.enderdrives.enderdisk.disabled": "§cThis EnderDisk is disabled on the server.",
+  "tooltip.enderdrives.fluid_types": "%s§7 of §9%s§7 Fluid Types",
+  "tooltip.enderdrives.fluidenderdisk.disabled": "§cThis FluidEnderDisk is disabled on the server.",
   "tooltip.enderdrives.frequency": "§7周波数: §e%s",
+  "tooltip.enderdrives.mode.bidirectional": "§7モード: §a双方向",
   "tooltip.enderdrives.mode.input": "§7モード: §9入力のみ",
   "tooltip.enderdrives.mode.output": "§7モード: §c出力のみ",
-  "tooltip.enderdrives.mode.bidirectional": "§7モード: §a双方向",
+  "tooltip.enderdrives.partitioned": "§7Partitioned: §f%s item%s",
+  "tooltip.enderdrives.scope.global": "§7グローバル",
   "tooltip.enderdrives.scope.private": "§7プライベート: §e%s",
   "tooltip.enderdrives.scope.team": "§7チーム: §e%s",
-  "tooltip.enderdrives.scope.global": "§7グローバル",
-  "tooltip.enderdrives.unknown": "不明",
+  "tooltip.enderdrives.tape.bytes": "%1$s の %2$s バイト使用済み",
   "tooltip.enderdrives.tape.types": "%1$s の %2$s タイプ",
-  "tooltip.enderdrives.tape.bytes": "%1$s の %2$s バイト使用済み"
+  "tooltip.enderdrives.types": "%s§7 の §9%s§7 タイプ",
+  "tooltip.enderdrives.unknown": "不明"
 }
