+++ common/src/main/resources/assets/enderdrives/lang/ru_ru.json
@@ -1,44 +1,59 @@
 {
-  "key.categories.enderdrives": "EnderDrives",
-  "item.enderdrives.ender_storage_component_1k": "1К компонент эндер хранения",
-  "item.enderdrives.ender_storage_component_4k": "4К компонент эндер хранения",
-  "item.enderdrives.ender_storage_component_16k": "16К компонент эндер хранения",
-  "item.enderdrives.ender_storage_component_64k": "64К компонент эндер хранения",
-  "item.enderdrives.ender_storage_component_256k": "256К компонент эндер хранения",
-  "item.enderdrives.ender_disk_1k": "1К предметная ячейка эндер хранения",
-  "item.enderdrives.ender_disk_4k": "4К предметная ячейка эндер хранения",
-  "item.enderdrives.ender_disk_16k": "16К предметная ячейка эндер хранения",
-  "item.enderdrives.ender_disk_64k": "64К предметная ячейка эндер хранения",
-  "item.enderdrives.ender_disk_256k": "256К предметная ячейка эндер хранения",
-  "item.enderdrives.ender_disk_creative": "Творческая предметная ячейка эндер хранения",
-  "item.enderdrives.tape_disk": "Дисковый накопитель эндер хранения",
   "enderdrives.transfer_mode.bidirectional": "Двунаправленный",
   "enderdrives.transfer_mode.input": "Только вход",
   "enderdrives.transfer_mode.output": "Только выход",
-  "screen.enderdrives.frequency_selector": "Селектор частот эндера",
+  "item.enderdrives.ender_disk_16k": "16К предметная ячейка эндер хранения",
+  "item.enderdrives.ender_disk_1k": "1К предметная ячейка эндер хранения",
+  "item.enderdrives.ender_disk_256k": "256К предметная ячейка эндер хранения",
+  "item.enderdrives.ender_disk_4k": "4К предметная ячейка эндер хранения",
+  "item.enderdrives.ender_disk_64k": "64К предметная ячейка эндер хранения",
+  "item.enderdrives.ender_disk_creative": "Творческая предметная ячейка эндер хранения",
+  "item.enderdrives.ender_fluid_disk_16k": "16k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_1k": "1k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_256k": "256k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_4k": "4k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_64k": "64k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_creative": "Creative Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_storage_component_16k": "16k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_1k": "1k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_256k": "256k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_4k": "4k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_64k": "64k Ender Fluid Storage Component",
+  "item.enderdrives.ender_storage_component_16k": "16К компонент эндер хранения",
+  "item.enderdrives.ender_storage_component_1k": "1К компонент эндер хранения",
+  "item.enderdrives.ender_storage_component_256k": "256К компонент эндер хранения",
+  "item.enderdrives.ender_storage_component_4k": "4К компонент эндер хранения",
+  "item.enderdrives.ender_storage_component_64k": "64К компонент эндер хранения",
+  "item.enderdrives.tape_disk": "Дисковый накопитель эндер хранения",
+  "key.categories.enderdrives": "EnderDrives",
   "screen.enderdrives.frequency": "Частота эндер-диска",
-  "screen.enderdrives.frequency_placeholder": "Частота...",
   "screen.enderdrives.frequency_label": "Частота",
-  "screen.enderdrives.toggle_scope": "Нажмите, чтобы переключить область действия",
-  "screen.enderdrives.tooltip.plus_256": "+256",
-  "screen.enderdrives.tooltip.plus_16": "+16",
-  "screen.enderdrives.tooltip.plus_1": "+1",
-  "screen.enderdrives.tooltip.minus_256": "-256",
-  "screen.enderdrives.tooltip.minus_16": "-16",
-  "screen.enderdrives.tooltip.minus_1": "-1",
+  "screen.enderdrives.frequency_placeholder": "Частота...",
+  "screen.enderdrives.frequency_selector": "Селектор частот эндера",
   "screen.enderdrives.scope.global": "Глобальный",
   "screen.enderdrives.scope.private": "Частный",
   "screen.enderdrives.scope.team": "Группа",
-  "tooltip.enderdrives.types": "%s§7 of §9%s§7 видов",
+  "screen.enderdrives.toggle_scope": "Нажмите, чтобы переключить область действия",
+  "screen.enderdrives.tooltip.minus_1": "-1",
+  "screen.enderdrives.tooltip.minus_16": "-16",
+  "screen.enderdrives.tooltip.minus_256": "-256",
+  "screen.enderdrives.tooltip.plus_1": "+1",
+  "screen.enderdrives.tooltip.plus_16": "+16",
+  "screen.enderdrives.tooltip.plus_256": "+256",
+  "tooltip.enderdrives.disabled": "§cThis item is disabled on the server.",
+  "tooltip.enderdrives.enderdisk.disabled": "§cThis EnderDisk is disabled on the server.",
+  "tooltip.enderdrives.fluid_types": "%s§7 of §9%s§7 Fluid Types",
+  "tooltip.enderdrives.fluidenderdisk.disabled": "§cThis FluidEnderDisk is disabled on the server.",
   "tooltip.enderdrives.frequency": "§7Частота: §e%s",
+  "tooltip.enderdrives.mode.bidirectional": "§7Режим: §aДвунаправленный",
   "tooltip.enderdrives.mode.input": "§7Режим: §9Только вход",
   "tooltip.enderdrives.mode.output": "§7Режим: §cТолько выход",
-  "tooltip.enderdrives.mode.bidirectional": "§7Режим: §aДвунаправленный",
+  "tooltip.enderdrives.partitioned": "§7Partitioned: §f%s item%s",
+  "tooltip.enderdrives.scope.global": "§7Глобальный",
   "tooltip.enderdrives.scope.private": "§7Частный: §e%s",
   "tooltip.enderdrives.scope.team": "§7Группа: §e%s",
-  "tooltip.enderdrives.scope.global": "§7Глобальный",
-  "tooltip.enderdrives.unknown": "Неизвестно",
-  "tooltip.enderdrives.tape.types": "%1$s из %2$s видов",
   "tooltip.enderdrives.tape.bytes": "%1$s из %2$s байт",
-  
+  "tooltip.enderdrives.tape.types": "%1$s из %2$s видов",
+  "tooltip.enderdrives.types": "%s§7 of §9%s§7 видов",
+  "tooltip.enderdrives.unknown": "Неизвестно"
 }
