+++ common/src/main/resources/assets/enderdrives/lang/zh_cn.json
@@ -1,45 +1,59 @@
 {
-  "key.categories.enderdrives": "末影驱动器",
-  "item.enderdrives.ender_storage_component_1k": "1k末影存储组件",
-  "item.enderdrives.ender_storage_component_4k": "4k末影存储组件",
-  "item.enderdrives.ender_storage_component_16k": "16k末影存储组件",
-  "item.enderdrives.ender_storage_component_64k": "64k末影存储组件",
-  "item.enderdrives.ender_storage_component_256k": "256k末影存储组件",
-  "item.enderdrives.ender_disk_1k": "1k末影物品存储元件",
-  "item.enderdrives.ender_disk_4k": "4k末影物品存储元件",
-  "item.enderdrives.ender_disk_16k": "16k末影物品存储元件",
-  "item.enderdrives.ender_disk_64k": "64k末影物品存储元件",
-  "item.enderdrives.ender_disk_256k": "256k末影物品存储元件",
-  "item.enderdrives.ender_disk_creative": "创造末影物品存储元件",
-  "item.enderdrives.tape_disk": "磁带盘物品存储元件",
   "enderdrives.transfer_mode.bidirectional": "双向",
   "enderdrives.transfer_mode.input": "仅输入",
   "enderdrives.transfer_mode.output": "仅输出",
-  "screen.enderdrives.frequency_selector": "末影频率选择器",
+  "item.enderdrives.ender_disk_16k": "16k末影物品存储元件",
+  "item.enderdrives.ender_disk_1k": "1k末影物品存储元件",
+  "item.enderdrives.ender_disk_256k": "256k末影物品存储元件",
+  "item.enderdrives.ender_disk_4k": "4k末影物品存储元件",
+  "item.enderdrives.ender_disk_64k": "64k末影物品存储元件",
+  "item.enderdrives.ender_disk_creative": "创造末影物品存储元件",
+  "item.enderdrives.ender_fluid_disk_16k": "16k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_1k": "1k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_256k": "256k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_4k": "4k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_64k": "64k Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_disk_creative": "Creative Ender Fluid Storage Cell",
+  "item.enderdrives.ender_fluid_storage_component_16k": "16k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_1k": "1k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_256k": "256k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_4k": "4k Ender Fluid Storage Component",
+  "item.enderdrives.ender_fluid_storage_component_64k": "64k Ender Fluid Storage Component",
+  "item.enderdrives.ender_storage_component_16k": "16k末影存储组件",
+  "item.enderdrives.ender_storage_component_1k": "1k末影存储组件",
+  "item.enderdrives.ender_storage_component_256k": "256k末影存储组件",
+  "item.enderdrives.ender_storage_component_4k": "4k末影存储组件",
+  "item.enderdrives.ender_storage_component_64k": "64k末影存储组件",
+  "item.enderdrives.tape_disk": "磁带盘物品存储元件",
+  "key.categories.enderdrives": "末影驱动器",
   "screen.enderdrives.frequency": "末影磁盘频率",
-  "screen.enderdrives.frequency_placeholder": "频率……",
   "screen.enderdrives.frequency_label": "频率",
-  "screen.enderdrives.toggle_scope": "点击以切换开放范围",
-  "screen.enderdrives.tooltip.plus_256": "+256",
-  "screen.enderdrives.tooltip.plus_16": "+16",
-  "screen.enderdrives.tooltip.plus_1": "+1",
-  "screen.enderdrives.tooltip.minus_256": "-256",
-  "screen.enderdrives.tooltip.minus_16": "-16",
-  "screen.enderdrives.tooltip.minus_1": "-1",
+  "screen.enderdrives.frequency_placeholder": "频率……",
+  "screen.enderdrives.frequency_selector": "末影频率选择器",
   "screen.enderdrives.scope.global": "公开",
   "screen.enderdrives.scope.private": "私人",
   "screen.enderdrives.scope.team": "团队",
-  "tooltip.enderdrives.types": "%s§7 / §9%s§7 类型",
+  "screen.enderdrives.toggle_scope": "点击以切换开放范围",
+  "screen.enderdrives.tooltip.minus_1": "-1",
+  "screen.enderdrives.tooltip.minus_16": "-16",
+  "screen.enderdrives.tooltip.minus_256": "-256",
+  "screen.enderdrives.tooltip.plus_1": "+1",
+  "screen.enderdrives.tooltip.plus_16": "+16",
+  "screen.enderdrives.tooltip.plus_256": "+256",
+  "tooltip.enderdrives.disabled": "§cThis item is disabled on the server.",
+  "tooltip.enderdrives.enderdisk.disabled": "§cThis EnderDisk is disabled on the server.",
+  "tooltip.enderdrives.fluid_types": "%s§7 of §9%s§7 Fluid Types",
+  "tooltip.enderdrives.fluidenderdisk.disabled": "§cThis FluidEnderDisk is disabled on the server.",
   "tooltip.enderdrives.frequency": "§7频率：§e%s",
+  "tooltip.enderdrives.mode.bidirectional": "§7模式：§a双向",
   "tooltip.enderdrives.mode.input": "§7模式：§9仅输入",
   "tooltip.enderdrives.mode.output": "§7模式：§c仅输出",
-  "tooltip.enderdrives.mode.bidirectional": "§7模式：§a双向",
+  "tooltip.enderdrives.partitioned": "§7已分区：§f%s个物品",
+  "tooltip.enderdrives.scope.global": "§7公开",
   "tooltip.enderdrives.scope.private": "§7私人：§e%s",
   "tooltip.enderdrives.scope.team": "§7团队：§e%s",
-  "tooltip.enderdrives.scope.global": "§7公开",
-  "tooltip.enderdrives.partitioned": "§7已分区：§f%s个物品",
-  "tooltip.enderdrives.unknown": "未知",
+  "tooltip.enderdrives.tape.bytes": "%1$s / %2$s 字节已使用",
   "tooltip.enderdrives.tape.types": "%1$s / %2$s 类型",
-  "tooltip.enderdrives.tape.bytes": "%1$s / %2$s 字节已使用"
-
+  "tooltip.enderdrives.types": "%s§7 / §9%s§7 类型",
+  "tooltip.enderdrives.unknown": "未知"
 }
